made BoardCard as a seperate component from viewboard page
